+++ Engine/Core/LuaScript/LuaScriptManager.h
@@ -61,11 +61,6 @@ public:
     T *loadConfig(const std::string &configName);
 
 private:
-    void initEvents();
-
-    void initLuaContext();
-
-    void loadLuaMain();
 
     template<typename TypeToRegister, typename... Members>
     void registerTypeImpl(refl::type_list<Members...>) noexcept;
