+++ LuaBind/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(LuaBind
         LuaBind.cpp
+        LuaBind_PreMain.cpp
 
         Config/LuaBind_Config.cpp
